Write out every item, even those without slots filled in.
It makes it much simpler for users who start by only identifying
the items they plan to use, and fill in the slots later.

It's also necessary; if this isn't done, a template reopened
may have only one craft or one drop instead of the intentional
drop + crafted.  I suspect it's why some users became confused.

Also it solves a problem with the linkage of outfits to itemid's.